Logging macros moved to the separate file
